--- edge-modules/PaidModuleSample/docker/linux/arm32v7/base/Dockerfile
@@ -1,6 +1,0 @@
-﻿ARG base_tag=3.1.18-bionic-arm32v7
-FROM mcr.microsoft.com/dotnet/runtime:${base_tag}
-
-# Add an unprivileged user account for running the module
-RUN useradd -ms /bin/bash moduleuser
-USER moduleuser
